Relocate generic_setattr into object protocol
As generic_setattr is included in object protocol, it seems more
appropriate to relocate.
